+++ src/app/Scenes/ArtQuiz/ArtQuizResults/ArtQuizResultsTabs/ArtQuizLikedArtworks.tsx
@@ -1,14 +1,20 @@
 import { ArtQuizLikedArtworks_me$key } from "__generated__/ArtQuizLikedArtworks_me.graphql"
-import { ArtQuizResultsQuery$data } from "__generated__/ArtQuizResultsQuery.graphql"
+import { ArtQuizResultsTabs_me$data } from "__generated__/ArtQuizResultsTabs_me.graphql"
 import GenericGrid from "app/Components/ArtworkGrids/GenericGrid"
 import { StickyTabPageScrollView } from "app/Components/StickyTabPage/StickyTabPageScrollView"
 import { useSpace } from "palette"
 import { graphql, useFragment } from "react-relay"
 import { useScreenDimensions } from "shared/hooks"
 
-export const ArtQuizLikedArtworks = ({ me }: { me: ArtQuizResultsQuery$data["me"] }) => {
-  const artworks = useFragment<ArtQuizLikedArtworks_me$key>(artQuizLikedArtworksFragment, me)?.quiz
-    .savedArtworks
+export const ArtQuizLikedArtworks = ({
+  savedArtworks,
+}: {
+  savedArtworks: ArtQuizResultsTabs_me$data["quiz"]["savedArtworks"]
+}) => {
+  const artworks = useFragment<ArtQuizLikedArtworks_me$key>(
+    artQuizLikedArtworksFragment,
+    savedArtworks
+  )
 
   const space = useSpace()
   const dimensions = useScreenDimensions()
@@ -20,7 +26,7 @@ export const ArtQuizLikedArtworks = ({ me }: { me: ArtQuizResultsQuery$data["me"
       }}
     >
       <GenericGrid
-        artworks={artworks!}
+        artworks={artworks}
         width={dimensions.width - space("2")}
         hidePartner
         artistNamesTextStyle={{ weight: "regular" }}
@@ -31,11 +37,7 @@ export const ArtQuizLikedArtworks = ({ me }: { me: ArtQuizResultsQuery$data["me"
 }
 
 const artQuizLikedArtworksFragment = graphql`
-  fragment ArtQuizLikedArtworks_me on Me {
-    quiz {
-      savedArtworks {
-        ...GenericGrid_artworks
-      }
-    }
+  fragment ArtQuizLikedArtworks_me on Artwork @relay(plural: true) {
+    ...GenericGrid_artworks
   }
 `
